APR/lab7/test-web/requirements.txt
alembic==1.18.4
annotated-types==0.7.0
blinker==1.9.0
click==8.3.3
dotenv==0.9.9
Flask==3.1.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.4.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.11
MarkupSafe==3.0.3
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic_core==2.46.3
python-dotenv==1.2.2
SQLAlchemy==2.0.49
typing-inspection==0.4.2
typing_extensions==4.15.0
Werkzeug==3.1.8
